requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
lxml>=4.9.0
schedule>=1.2.0
python-dateutil>=2.8.0
jinja2>=3.1.0
feedparser>=6.0.0
